Changed hash_table_t to use size_t instead of int where appropriate
* size can never be greater than size_t, so it makes sense to use size_t
for it.
* next_split should always point to an array element we can address
directly, so it makes sense to use size_t for it as well.
* While n_elements can be greater than size in theory, as it is now it
will cause the problem as we try to grow the table, so for now I'm going
with size_t for it as well.

+++ hashtable.h
@@ -1,6 +1,8 @@
 #ifndef HASTABLE_H_
 #define HASHTABLE_H_
 
+#include <stddef.h>
+
 //Increase the size of the table when the ratio between the number of stored
 //elements and the number of slots is greater than this.
 #define HASH_TABLE_GROW_SIZE 0.75
@@ -14,10 +16,13 @@ typedef struct _hash_table_list {
 } hash_table_list_t;
 
 typedef struct _hash_table {
-	int size;
-	int n_elements;
+	size_t size;
+	//While n_elements potentially could be greater than size_t, this code will
+	//not work properly if you try to grow the table further, so having a
+	//larger n_elements makes no sense.
+	size_t n_elements;
 	int exponent;
-	int next_split;
+	size_t next_split;
 	hash_table_list_t **elements;
 } hash_table_t;
